requirements.txt
docx2txt==0.9
jupyterlab==4.4.2
llama-index-agent-openai==0.4.7
llama-index-core==0.12.35
llama-index-embeddings-openai==0.3.1
llama-index-llms-openai==0.3.38
llama-index-readers-file==0.4.7
llama-index-readers-google==0.6.1
llama-index-vector-stores-qdrant==0.6.0
langchain-community==0.3.24
mypy==1.15.0
python-dotenv==1.1.0
qdrant-client==1.14.2
types-requests